+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Upcoming Version
+
+**Added:**
+- Azure SQL Service - Added support for SQL Server Firewall Rules.
+
 ## 0.0.9
 
 **Added:**
@@ -5,8 +10,6 @@
 
 **Changed:**
 - DNS Service - DNS Service moved from API calls to Azure latest SDK.
-
-
 
 ## 0.0.8 
 
@@ -19,7 +22,6 @@
     get blob properties, set blob metadata, get blob metadata, get blob, list blob functionality for storage
 - Added get container properties, get container metadata, set container metadata, get container, list container,
     delete container,  Get the access control list of the container
-
 
 ## 0.0.6